requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
docx2txt
pdfminer.six
scikit-learn
nltk
